Validate individual blocks from an upstream component
When blocks come from a peer consumers should check to see if the blocks
are valid before doing anything else with them. This stream checks to
see if the parent_id of a block matches the previous block, that the
block hash is correct, and that optionally, the ledger database
doesn't already contain block being passed.